refactor: 패키지 구조 변경 및 미사용 클래스 삭제
