+++ src/main/java/chess/controller/dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.controller.dto;
 
 public class ResponseDto {
     private final String message;
