+++ src/main/java/chess/domain/dao/BoardJdbcTemplateDao.java
@@ -1,6 +1,6 @@
 package chess.domain.dao;
 
-import chess.domain.dto.PieceDto;
+import chess.service.dto.PieceDto;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
