+++ src/main/java/chess/service/dto/PieceDto.java
@@ -1,4 +1,4 @@
-package chess.domain.dto;
+package chess.service.dto;
 
 public class PieceDto {
